requirements.txt
anthropic
python-dotenv
requests
pillow 